+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Match graphics engine framerate with source content (or channel normalized) framerate
   - Fix loading requested number of epg entries for motion graphics elements
 - Fix bug with mirror channels where seemingly random content would be played every ~40 seconds
+- Fix chronological sorting for Other Videos that have release date metadata
 
 ### Changed
 - No longer round framerate to nearest integer when normalizing framerate